Move functions to vehicle 1
