+++ src/OpenLoco/src/Vehicles/Vehicle.cpp
@@ -360,49 +360,6 @@ namespace OpenLoco::Vehicles
         return false;
     }
 
-    static constexpr uint8_t getMovementNibble(const World::Pos3& pos1, const World::Pos3& pos2)
-    {
-        uint8_t nibble = 0;
-        if (pos1.x != pos2.x)
-        {
-            nibble |= (1U << 0);
-        }
-        if (pos1.y != pos2.y)
-        {
-            nibble |= (1U << 1);
-        }
-        if (pos1.z != pos2.z)
-        {
-            nibble |= (1U << 2);
-        }
-        return nibble;
-    }
-
-    // 0x00500120
-    static constexpr std::array<uint32_t, 8> movementNibbleToDistance = {
-        0,
-        0x220C,
-        0x220C,
-        0x3027,
-        0x199A,
-        0x2A99,
-        0x2A99,
-        0x3689,
-    };
-
-    // 0x00500244
-    static constexpr std::array<World::TilePos2, 9> kMooreNeighbourhood = {
-        World::TilePos2{ 0, 0 },
-        World::TilePos2{ 0, 1 },
-        World::TilePos2{ 1, 1 },
-        World::TilePos2{ 1, 0 },
-        World::TilePos2{ 1, -1 },
-        World::TilePos2{ 0, -1 },
-        World::TilePos2{ -1, -1 },
-        World::TilePos2{ -1, 0 },
-        World::TilePos2{ -1, 1 },
-    };
-
     // If candidate within 8 vehicle components of src we ignore a self collision
     // TODO: If we stored the car index this could be simplified
     static bool ignoreSelfCollision(VehicleBase& sourceVehicleId, const VehicleBase& candidateVehicleId)
@@ -493,439 +450,6 @@ namespace OpenLoco::Vehicles
         return EntityId::null;
     }
 
-    enum class OvertakeResult
-    {
-        overtakeAvailble,
-        noOvertakeAvailble,
-        mayBeOvertaken,
-    };
-
-    // 0x0047CD78
-    // ax : pos.x
-    // cx : pos.y
-    // dl : pos.z / World::kSmallZStep
-    // ebp : tad
-    // esi : veh1
-    //
-    // return mayBeOvertaken == high carry flag
-    //        overtakeAvailable   0x0112C328 == 5
-    //        noOvertakeAvailable 0x0112C328 == 0
-    static OvertakeResult getRoadOvertakeAvailability(const Vehicle1& veh1, World::Pos3 pos, uint16_t tad)
-    {
-        OvertakeResult result = OvertakeResult::noOvertakeAvailble;
-        for (const auto& nearby : kMooreNeighbourhood)
-        {
-            const auto inspectionPos = World::toTileSpace(pos) + nearby;
-            for (auto* entity : EntityManager::EntityTileList(World::toWorldSpace(inspectionPos)))
-            {
-                auto* vehicleBase = entity->asBase<VehicleBase>();
-                if (vehicleBase == nullptr || !vehicleBase->isVehicleTail())
-                {
-                    continue;
-                }
-                auto* vehicleTail = vehicleBase->asVehicleTail();
-                if (vehicleTail == nullptr)
-                {
-                    continue;
-                }
-                if (vehicleTail->getTransportMode() != TransportMode::road)
-                {
-                    continue;
-                }
-                if (vehicleTail->tileBaseZ * World::kSmallZStep != pos.z)
-                {
-                    continue;
-                }
-                if (vehicleTail->tileX != pos.x || vehicleTail->tileY != pos.y)
-                {
-                    continue;
-                }
-                if ((vehicleTail->trackAndDirection.road._data & World::Track::AdditionalTaDFlags::basicTaDMask) != (tad & World::Track::AdditionalTaDFlags::basicTaDMask))
-                {
-                    continue;
-                }
-
-                // Perhaps a little expensive to do this (we don't do it on rail vehicles for example)
-                Vehicle train(vehicleTail->head);
-                if (train.veh1->var_3C < 0x220C0)
-                {
-                    continue;
-                }
-                if ((train.veh2->var_73 & Flags73::isBrokenDown) != Flags73::none)
-                {
-                    continue;
-                }
-                if (veh1.var_3C < train.veh1->var_3C)
-                {
-                    return OvertakeResult::mayBeOvertaken;
-                }
-                auto* veh2 = EntityManager::get<Vehicle2>(veh1.nextCarId);
-                if (veh2 == nullptr)
-                {
-                    continue;
-                }
-                if (veh2->maxSpeed <= train.veh2->maxSpeed)
-                {
-                    return OvertakeResult::mayBeOvertaken;
-                }
-                result = OvertakeResult::overtakeAvailble;
-            }
-        }
-        return result;
-    }
-
-    struct Sub47CEB7Result
-    {
-        uint8_t al;
-        uint8_t ah;
-    };
-
-    // 0x0047CEB7
-    // veh1 : esi
-    // return : al, ah
-    static Sub47CEB7Result sub_47CEB7(Vehicle1& veh1)
-    {
-        const auto startPos = World::Pos3(veh1.tileX, veh1.tileY, veh1.tileBaseZ * World::kSmallZStep);
-        uint8_t unk112C327 = 0;
-
-        auto pos = startPos + World::TrackData::getUnkRoad(veh1.trackAndDirection.road._data & 0x7F).pos;
-        unk112C327++;
-
-        auto routings = RoutingManager::RingView(veh1.routingHandle);
-        auto routingIter = routings.begin();
-        if (routingIter == routings.end())
-        {
-            return Sub47CEB7Result{ 0, 0 };
-        }
-        routingIter++;
-        if (routingIter == routings.end())
-        {
-            return Sub47CEB7Result{ 0, 0 };
-        }
-        auto routing = RoutingManager::getRouting(*routingIter);
-        TrackAndDirection::_RoadAndDirection tad(0, 0);
-        tad._data = routing & World::Track::AdditionalTaDFlags::basicTaDMask;
-        const auto occupationFlags = getRoadOccupation(pos, tad);
-
-        if ((occupationFlags & RoadOccupationFlags::hasLevelCrossing) != RoadOccupationFlags::none)
-        {
-            // 0x0047D0B9
-            if ((occupationFlags & (RoadOccupationFlags::isLaneOccupied | RoadOccupationFlags::isLevelCrossingClosed)) != RoadOccupationFlags::none)
-            {
-                return Sub47CEB7Result{ 2, 0 };
-            }
-            ++unk112C327;
-            for (; unk112C327 < 255; ++unk112C327)
-            {
-                pos += World::TrackData::getUnkRoad(tad._data & 0x7F).pos;
-                routingIter++;
-                if (routingIter == routings.end())
-                {
-                    return Sub47CEB7Result{ 2, 0 };
-                }
-                tad._data = RoutingManager::getRouting(*routingIter) & World::Track::AdditionalTaDFlags::basicTaDMask;
-
-                const auto fwdOccupationFlags = getRoadOccupation(pos, tad);
-                if ((fwdOccupationFlags & (RoadOccupationFlags::isLaneOccupied | RoadOccupationFlags::isLevelCrossingClosed)) != RoadOccupationFlags::none)
-                {
-                    return Sub47CEB7Result{ 2, 0 };
-                }
-                if ((fwdOccupationFlags & RoadOccupationFlags::hasLevelCrossing) == RoadOccupationFlags::none)
-                {
-                    return Sub47CEB7Result{ 4, unk112C327 };
-                }
-            }
-            return Sub47CEB7Result{ 2, 0 };
-        }
-        else
-        {
-            OvertakeResult overtakeResult = OvertakeResult::overtakeAvailble;
-            if ((occupationFlags & RoadOccupationFlags::isLaneOccupied) != RoadOccupationFlags::none)
-            {
-                overtakeResult = getRoadOvertakeAvailability(veh1, pos, tad._data);
-                if (overtakeResult == OvertakeResult::mayBeOvertaken)
-                {
-                    return Sub47CEB7Result{ 0, 0 };
-                }
-            }
-            else
-            {
-                // 0x0047CF32
-                auto nextRoutingIter = routingIter;
-                nextRoutingIter++;
-                const auto nextPos = pos + World::TrackData::getUnkRoad(tad._data & 0x7F).pos;
-                if (nextRoutingIter == routings.end())
-                {
-                    return Sub47CEB7Result{ 0, 0 };
-                }
-                auto nextRouting = RoutingManager::getRouting(*nextRoutingIter);
-                TrackAndDirection::_RoadAndDirection nextTad(0, 0);
-                nextTad._data = nextRouting & World::Track::AdditionalTaDFlags::basicTaDMask;
-                const auto nextOccupationFlags = getRoadOccupation(nextPos, nextTad);
-                if ((nextOccupationFlags & RoadOccupationFlags::isLaneOccupied) == RoadOccupationFlags::none)
-                {
-                    return Sub47CEB7Result{ 0, 0 };
-                }
-                overtakeResult = getRoadOvertakeAvailability(veh1, nextPos, nextTad._data);
-                if (overtakeResult == OvertakeResult::mayBeOvertaken)
-                {
-                    return Sub47CEB7Result{ 0, 0 };
-                }
-            }
-            // 0x0047CFB5
-            const auto& roadObj = ObjectManager::get<RoadObject>(veh1.trackType);
-            if (!roadObj->hasFlags(RoadObjectFlags::isRoad))
-            {
-                return Sub47CEB7Result{ 0, 0 };
-            }
-            if (tad.isUnk8() || tad.isBackToFront())
-            {
-                return Sub47CEB7Result{ 0, 0 };
-            }
-            auto backToFrontTad = tad;
-            backToFrontTad._data ^= (1U << 7);
-            auto btfOccupationFlags = getRoadOccupation(pos, backToFrontTad);
-            if ((btfOccupationFlags & (RoadOccupationFlags::hasLevelCrossing | RoadOccupationFlags::isLaneOccupied)) != RoadOccupationFlags::none)
-            {
-                return Sub47CEB7Result{ 0, 0 };
-            }
-            const auto unkCount = overtakeResult == OvertakeResult::overtakeAvailble ? 5 : 0;
-            ++unk112C327;
-            auto i = 0;
-            for (; unk112C327 < 11; ++unk112C327)
-            {
-                pos += World::TrackData::getUnkRoad(tad._data & 0x7F).pos;
-                routingIter++;
-                if (routingIter == routings.end())
-                {
-                    if ((btfOccupationFlags & RoadOccupationFlags::hasStation) == RoadOccupationFlags::none)
-                    {
-                        return Sub47CEB7Result{ 0, 0 };
-                    }
-                    return Sub47CEB7Result{ 8, static_cast<uint8_t>(unk112C327 - 1) };
-                }
-                tad._data = RoutingManager::getRouting(*routingIter) & World::Track::AdditionalTaDFlags::basicTaDMask;
-
-                backToFrontTad = tad;
-                backToFrontTad._data ^= (1U << 7);
-                btfOccupationFlags = getRoadOccupation(pos, backToFrontTad);
-                if ((btfOccupationFlags & (RoadOccupationFlags::hasLevelCrossing | RoadOccupationFlags::isLaneOccupied)) != RoadOccupationFlags::none)
-                {
-                    return Sub47CEB7Result{ 0, 0 };
-                }
-                const auto fwdOccupationFlags = getRoadOccupation(pos, tad);
-                if ((fwdOccupationFlags & RoadOccupationFlags::isLaneOccupied) != RoadOccupationFlags::none)
-                {
-                    i = 0;
-                    continue;
-                }
-                i++;
-                if (i <= unkCount)
-                {
-                    continue;
-                }
-                return Sub47CEB7Result{ 1, unk112C327 };
-            }
-            return Sub47CEB7Result{ 0, 0 };
-        }
-    }
-
-    // 0x0047D308
-    static bool veh1UpdateRoadMotionNewRoadPiece(Vehicle1& veh1)
-    {
-        auto newRoutingHandle = veh1.routingHandle;
-        auto newIndex = newRoutingHandle.getIndex() + 1;
-        newRoutingHandle.setIndex(newIndex);
-        const auto routing = RoutingManager::getRouting(newRoutingHandle);
-        if (routing == RoutingManager::kAllocatedButFreeRoutingStation)
-        {
-            return false;
-        }
-
-        const auto newPos = World::Pos3(veh1.tileX, veh1.tileY, veh1.tileBaseZ * World::kSmallZStep)
-            + World::TrackData::getUnkRoad(veh1.trackAndDirection.road._data & 0x7F).pos;
-
-        TrackAndDirection::_RoadAndDirection newRad(0, 0);
-        newRad._data = routing & 0x1FFU;
-        veh1.sub_47D959(newPos, newRad, true);
-
-        veh1.routingHandle = newRoutingHandle;
-        veh1.trackAndDirection.road = newRad;
-
-        veh1.tileX = newPos.x;
-        veh1.tileY = newPos.y;
-        veh1.tileBaseZ = newPos.z / World::kSmallZStep;
-        return true;
-    }
-
-    // 0x0047D2D6
-    // veh1 : esi
-    // numRoadPieces : ah
-    // return eax
-    static int32_t vehicle1UpdateRoadMotionByPieces(Vehicle1& veh1, const uint8_t numRoadPieces)
-    {
-        auto intermediatePosition = veh1.position;
-        auto distanceMoved = 0;
-        for (auto i = 0; i < numRoadPieces;)
-        {
-            auto newSubPosition = veh1.subPosition + 1U;
-            const auto subPositionDataSize = World::TrackData::getRoadSubPositon(veh1.trackAndDirection.road._data).size();
-            // This means we have moved forward by a road piece
-            if (newSubPosition >= subPositionDataSize)
-            {
-                if (!veh1UpdateRoadMotionNewRoadPiece(veh1))
-                {
-                    // This seems wrong but its what vanilla does
-                    // Todo investigate if we should return 0 or actually move
-                    // the vehicle
-                    return distanceMoved;
-                }
-                else
-                {
-                    newSubPosition = 0;
-                    i++;
-                }
-            }
-            veh1.subPosition = newSubPosition;
-            const auto& moveData = World::TrackData::getRoadSubPositon(veh1.trackAndDirection.road._data)[newSubPosition];
-            const auto nextNewPosition = moveData.loc + World::Pos3(veh1.tileX, veh1.tileY, veh1.tileBaseZ * World::kSmallZStep);
-            distanceMoved += movementNibbleToDistance[getMovementNibble(intermediatePosition, nextNewPosition)];
-            intermediatePosition = nextNewPosition;
-            veh1.spriteYaw = moveData.yaw;
-            veh1.spritePitch = moveData.pitch;
-        }
-
-        veh1.moveTo(intermediatePosition);
-        return distanceMoved;
-    }
-
-    // 0x0047D1B4
-    // Similar to the others but doesn't update the vehicle
-    // due to that it must take reference to all the variables
-    static bool veh1UpdateRoadMotionNewRoadPieceNoMove(World::Pos3& pos, RoutingHandle& handle, TrackAndDirection::_RoadAndDirection& rad)
-    {
-        auto newIndex = handle.getIndex() + 1;
-        handle.setIndex(newIndex);
-        const auto routing = RoutingManager::getRouting(handle);
-        if (routing == RoutingManager::kAllocatedButFreeRoutingStation)
-        {
-            return false;
-        }
-
-        pos += World::TrackData::getUnkRoad(rad._data & 0x7F).pos;
-
-        rad._data = routing & 0x1FFU;
-        return true;
-    }
-
-    // 0x0047D142
-    // veh1 : esi
-    // numRoadPieces : ah
-    // return eax
-    static int32_t vehicle1UpdateRoadMotionByPiecesNoMove(Vehicle1& veh1, const uint8_t numRoadPieces)
-    {
-        auto intermediatePosition = veh1.position;
-        auto distanceMoved = 0;
-        auto pos = World::Pos3(veh1.tileX, veh1.tileY, veh1.tileBaseZ * World::kSmallZStep);
-        auto rad = veh1.trackAndDirection.road;
-        auto handle = veh1.routingHandle;
-        auto subPosition = veh1.subPosition;
-        for (auto i = 0; i < numRoadPieces;)
-        {
-            auto newSubPosition = subPosition + 1U;
-            const auto subPositionDataSize = World::TrackData::getRoadSubPositon(rad._data).size();
-            // This means we have moved forward by a road piece
-            if (newSubPosition >= subPositionDataSize)
-            {
-                if (!veh1UpdateRoadMotionNewRoadPieceNoMove(pos, handle, rad))
-                {
-                    return distanceMoved;
-                }
-                else
-                {
-                    newSubPosition = 0;
-                    i++;
-                }
-            }
-            subPosition = newSubPosition;
-            const auto& moveData = World::TrackData::getRoadSubPositon(rad._data)[newSubPosition];
-            const auto nextNewPosition = moveData.loc + pos;
-            distanceMoved += movementNibbleToDistance[getMovementNibble(intermediatePosition, nextNewPosition)];
-            intermediatePosition = nextNewPosition;
-        }
-        return distanceMoved;
-    }
-
-    // 0x0047D46F
-    // veh1 : esi
-    // numRoadPieces : ah
-    static void applyOvertakeToVehicle1(Vehicle1& veh1, uint8_t numRoadPieces)
-    {
-        const auto noOvertakeDistance = vehicle1UpdateRoadMotionByPiecesNoMove(veh1, numRoadPieces);
-        auto routingRing = RoutingManager::RingView(veh1.routingHandle);
-        auto iter = ++routingRing.begin();
-        constexpr auto kResetRouting = static_cast<uint16_t>(~((1U << 7) | (1U << 8)));
-        for (auto i = 0; i < numRoadPieces; ++iter, ++i)
-        {
-            if (iter == routingRing.end())
-            {
-                break;
-            }
-            const auto routing = RoutingManager::getRouting(*iter);
-            if (i == 0)
-            {
-                RoutingManager::setRouting(*iter, (routing & kResetRouting) | (1U << 8));
-            }
-            else if (i == (numRoadPieces - 1))
-            {
-                RoutingManager::setRouting(*iter, (routing & kResetRouting) | (1U << 8) | (1U << 7));
-            }
-            else
-            {
-                RoutingManager::setRouting(*iter, (routing & kResetRouting) | (1 << 7));
-            }
-        }
-
-        const auto newDistance = vehicle1UpdateRoadMotionByPieces(veh1, numRoadPieces);
-        veh1.var_3C += newDistance;
-        auto* head = EntityManager::get<VehicleHead>(veh1.head);
-        head->var_3C += newDistance - noOvertakeDistance;
-    }
-
-    // 0x0047D52B
-    // veh1 : esi
-    // numRoadPieces : ah
-    static void applyChangeLaneToVehicle1(Vehicle1& veh1, uint8_t numRoadPieces)
-    {
-        const auto noOvertakeDistance = vehicle1UpdateRoadMotionByPiecesNoMove(veh1, numRoadPieces);
-        auto routingRing = RoutingManager::RingView(veh1.routingHandle);
-        auto iter = ++routingRing.begin();
-        constexpr auto kResetRouting = static_cast<uint16_t>(~((1U << 7) | (1U << 8)));
-        for (auto i = 0; i < numRoadPieces; ++iter, ++i)
-        {
-            if (iter == routingRing.end())
-            {
-                break;
-            }
-            const auto routing = RoutingManager::getRouting(*iter);
-            if (i == 0)
-            {
-                RoutingManager::setRouting(*iter, (routing & kResetRouting) | (1U << 8));
-            }
-            else
-            {
-                RoutingManager::setRouting(*iter, (routing & kResetRouting) | (1 << 7));
-            }
-        }
-
-        const auto newDistance = vehicle1UpdateRoadMotionByPieces(veh1, numRoadPieces);
-        veh1.var_3C += newDistance;
-        auto* head = EntityManager::get<VehicleHead>(veh1.head);
-        head->var_3C += newDistance - noOvertakeDistance;
-        head->trackAndDirection.road._data &= kResetRouting;
-        head->trackAndDirection.road._data |= numRoadPieces <= 1 ? (1U << 8) : (1U << 7);
-    }
-
     // 0x0047C7FA
     static int32_t updateRoadMotion(VehicleCommon& component, int32_t distance)
     {
@@ -957,7 +481,7 @@ namespace OpenLoco::Vehicles
             component.subPosition = newSubPosition;
             const auto& moveData = World::TrackData::getRoadSubPositon(component.trackAndDirection.road._data)[newSubPosition];
             const auto nextNewPosition = moveData.loc + World::Pos3(component.tileX, component.tileY, component.tileBaseZ * World::kSmallZStep);
-            component.remainingDistance -= movementNibbleToDistance[getMovementNibble(intermediatePosition, nextNewPosition)];
+            component.remainingDistance -= kMovementNibbleToDistance[getMovementNibble(intermediatePosition, nextNewPosition)];
             intermediatePosition = nextNewPosition;
             component.spriteYaw = moveData.yaw;
             component.spritePitch = moveData.pitch;
@@ -1017,7 +541,7 @@ namespace OpenLoco::Vehicles
                 component.subPosition = newSubPosition;
                 const auto& moveData = World::TrackData::getTrackSubPositon(component.trackAndDirection.track._data)[newSubPosition];
                 const auto nextNewPosition = moveData.loc + World::Pos3(component.tileX, component.tileY, component.tileBaseZ * World::kSmallZStep);
-                component.remainingDistance -= movementNibbleToDistance[getMovementNibble(intermediatePosition, nextNewPosition)];
+                component.remainingDistance -= kMovementNibbleToDistance[getMovementNibble(intermediatePosition, nextNewPosition)];
                 intermediatePosition = nextNewPosition;
                 component.spriteYaw = moveData.yaw;
                 component.spritePitch = moveData.pitch;
@@ -1650,90 +1174,12 @@ namespace OpenLoco::Vehicles
             });
 
         registerHook(
-            0x0047CD78,
-            [](registers& regs) FORCE_ALIGN_ARG_POINTER -> uint8_t {
-                registers backup = regs;
-
-                Vehicle1* veh1 = X86Pointer<Vehicle1>(regs.esi);
-                const auto pos = World::Pos3(regs.ax, regs.cx, regs.dl * World::kSmallZStep);
-                const uint16_t tad = regs.bp;
-
-                const auto res = getRoadOvertakeAvailability(*veh1, pos, tad);
-
-                addr<0x0112C328, uint8_t>() = res == OvertakeResult::overtakeAvailble ? 5 : 0;
-
-                regs = backup;
-                return res == OvertakeResult::mayBeOvertaken ? X86_FLAG_CARRY : 0;
-            });
-
-        registerHook(
-            0x0047CEB7,
-            [](registers& regs) FORCE_ALIGN_ARG_POINTER -> uint8_t {
-                registers backup = regs;
-
-                Vehicle1* veh1 = X86Pointer<Vehicle1>(regs.esi);
-
-                const auto res = sub_47CEB7(*veh1);
-
-                regs = backup;
-                regs.al = res.al;
-                regs.ah = res.ah;
-                return 0;
-            });
-
-        registerHook(
-            0x0047D2D6,
-            [](registers& regs) FORCE_ALIGN_ARG_POINTER -> uint8_t {
-                registers backup = regs;
-                Vehicle1* veh1 = X86Pointer<Vehicle1>(regs.esi);
-                const uint8_t numRoadPieces = regs.ah;
-                const auto res = vehicle1UpdateRoadMotionByPieces(*veh1, numRoadPieces);
-                regs = backup;
-                regs.eax = res;
-                return 0;
-            });
-
-        registerHook(
-            0x0047D142,
-            [](registers& regs) FORCE_ALIGN_ARG_POINTER -> uint8_t {
-                registers backup = regs;
-                Vehicle1* veh1 = X86Pointer<Vehicle1>(regs.esi);
-                const uint8_t numRoadPieces = regs.ah;
-                const auto res = vehicle1UpdateRoadMotionByPiecesNoMove(*veh1, numRoadPieces);
-                regs = backup;
-                regs.eax = res;
-                return 0;
-            });
-
-        registerHook(
             0x004AE2AB,
             [](registers& regs) FORCE_ALIGN_ARG_POINTER -> uint8_t {
                 registers backup = regs;
                 VehicleHead* head = X86Pointer<VehicleHead>(regs.esi);
                 Vehicle train(*head);
                 applyVehicleObjectLength(train);
-                regs = backup;
-                return 0;
-            });
-
-        registerHook(
-            0x0047D46F,
-            [](registers& regs) FORCE_ALIGN_ARG_POINTER -> uint8_t {
-                registers backup = regs;
-                Vehicle1* veh1 = X86Pointer<Vehicle1>(regs.esi);
-                const uint8_t numRoadPieces = regs.ah;
-                applyOvertakeToVehicle1(*veh1, numRoadPieces);
-                regs = backup;
-                return 0;
-            });
-
-        registerHook(
-            0x0047D52B,
-            [](registers& regs) FORCE_ALIGN_ARG_POINTER -> uint8_t {
-                registers backup = regs;
-                Vehicle1* veh1 = X86Pointer<Vehicle1>(regs.esi);
-                const uint8_t numRoadPieces = regs.ah;
-                applyChangeLaneToVehicle1(*veh1, numRoadPieces);
                 regs = backup;
                 return 0;
             });
