+++ src/OpenLoco/src/Vehicles/Vehicle.h
@@ -229,6 +229,49 @@ namespace OpenLoco::Vehicles
     };
     OPENLOCO_ENABLE_ENUM_OPERATORS(SignalStateFlags);
 
+    static constexpr uint8_t getMovementNibble(const World::Pos3& pos1, const World::Pos3& pos2)
+    {
+        uint8_t nibble = 0;
+        if (pos1.x != pos2.x)
+        {
+            nibble |= (1U << 0);
+        }
+        if (pos1.y != pos2.y)
+        {
+            nibble |= (1U << 1);
+        }
+        if (pos1.z != pos2.z)
+        {
+            nibble |= (1U << 2);
+        }
+        return nibble;
+    }
+
+    // 0x00500120
+    static constexpr std::array<uint32_t, 8> kMovementNibbleToDistance = {
+        0,
+        0x220C,
+        0x220C,
+        0x3027,
+        0x199A,
+        0x2A99,
+        0x2A99,
+        0x3689,
+    };
+
+    // 0x00500244
+    static constexpr std::array<World::TilePos2, 9> kMooreNeighbourhood = {
+        World::TilePos2{ 0, 0 },
+        World::TilePos2{ 0, 1 },
+        World::TilePos2{ 1, 1 },
+        World::TilePos2{ 1, 0 },
+        World::TilePos2{ 1, -1 },
+        World::TilePos2{ 0, -1 },
+        World::TilePos2{ -1, -1 },
+        World::TilePos2{ -1, 0 },
+        World::TilePos2{ -1, 1 },
+    };
+
     void setSignalState(const World::Pos3& loc, const TrackAndDirection::_TrackAndDirection trackAndDirection, const uint8_t trackType, uint32_t flags);
     SignalStateFlags getSignalState(const World::Pos3& loc, const TrackAndDirection::_TrackAndDirection trackAndDirection, const uint8_t trackType, uint32_t flags);
     void sub_4A2AD7(const World::Pos3& loc, const TrackAndDirection::_TrackAndDirection trackAndDirection, const CompanyId company, const uint8_t trackType);
